Create Set-ADUser-HomeDrives.ps1
Creates home directory based on user SAM name and sets permissions to "user/modify" with inheritance and gives ownership to user. If folder already exists, permissions will be changed to "user/modify" with inheritance but will not force propagation. If user already has a home path set that does not meet the script configuration, that path will be changed.